backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
uvicorn==0.34.3
starlette==0.46.2
fastapi==0.115.12
resend==2.23.0
